Added Custom error pages (400, 404, 500)

+++ Vitalis/Vitalis/Controllers/CatalogController.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Vitalis.Data;
+using Vitalis.Data.Models;
 using Vitalis.Services.Core;
 using Vitalis.Services.Core.Contracts;
 using Vitalis.Web.Controllers;
@@ -21,22 +22,26 @@ namespace Vitalis.Controllers
         
         [HttpGet]
         [AllowAnonymous]
-        public async Task<IActionResult> Meals(string? searchQuery = null)
+        public async Task<IActionResult> Meals(string? searchQuery = null, int pageNumber = 1, int pageSize = 9)
         {
-            IEnumerable<MealViewModel> meals = await catalogService.GetAllMealsAsync(searchQuery);
+            var (meals, totalPages) = await catalogService.GetAllMealsAsync(searchQuery, pageNumber, pageSize);
 
             ViewData["SearchQuery"] = searchQuery;
+            ViewData["CurrentPage"] = pageNumber;
+            ViewData["TotalPages"] = totalPages;
             return View(meals);
         }
 
         
         [HttpGet]
         [AllowAnonymous]
-        public async Task<IActionResult> Ingredients(string? searchQuery = null)
+        public async Task<IActionResult> Ingredients(string? searchQuery = null, int pageNumber = 1, int pageSize = 9)
         {
-            IEnumerable<IngredientViewModel> ingredients = await catalogService.GetAllIngredientsAsync(searchQuery);
+            var(ingredients, totalPages) = await catalogService.GetAllIngredientsAsync(searchQuery, pageNumber, pageSize);
 
             ViewData["SearchQuery"] = searchQuery;
+            ViewData["CurrentPage"] = pageNumber;
+            ViewData["TotalPages"] = totalPages;
             return View(ingredients);
              
         }
@@ -54,7 +59,15 @@ namespace Vitalis.Controllers
         [AllowAnonymous]
         public async Task<IActionResult> ViewMeal(int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest();
+            }
             MealViewModel meal = await catalogService.GetMealByIdAsync(id);
+            if (meal == null)
+            {
+                return NotFound();
+            }
             return View(meal);
         }
 
@@ -62,15 +75,26 @@ namespace Vitalis.Controllers
         [AllowAnonymous]
         public async Task<IActionResult> ViewIngredient(int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest();
+            }
             IngredientViewModel ingredient = await catalogService.GetIngredientByIdAsync(id);
+            if (ingredient == null)
+            {
+                return NotFound();
+            }
             return View(ingredient);
         }
 
         [HttpPost]
         public async Task<IActionResult> DeleteMeal(int id, int TagId)
         {
+            if (id <= 0)
+            {
+                return BadRequest();
+            }
             await catalogService.DeleteMealAsync(id);
-
             if (TagId != 0) return RedirectToAction("ViewTag", new { id =TagId });
             else return RedirectToAction("Meals");
         }
@@ -80,6 +104,10 @@ namespace Vitalis.Controllers
         [Authorize(Roles ="Admin")]
         public async Task<IActionResult> DeleteIngredient(int id, int TagId)
         {
+            if (id <= 0)
+            {
+                return BadRequest();
+            }
             await catalogService.DeleteIngredientAsync(id);
 
             if (TagId != 0) return RedirectToAction("ViewTag", new { id = TagId });
@@ -91,6 +119,10 @@ namespace Vitalis.Controllers
         [Authorize(Roles = "Admin")]
         public async Task<IActionResult> DeleteTag(int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest();
+            }
             await catalogService.DeleteTagAsync(id);
 
             return RedirectToAction("Tags", "Catalog");
@@ -100,8 +132,15 @@ namespace Vitalis.Controllers
         [AllowAnonymous]
         public async Task<IActionResult> ViewTag(int id)
         {
+            if (id <= 0)
+            {
+                return BadRequest();
+            }
             ViewTagViewModel tagview = await catalogService.GetViewByTagAsync(id);
-
+            if (tagview == null)
+            {
+                return NotFound();
+            }
             return View(tagview);
         }
